requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
docutils==0.14
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
lockfile==0.12.2
luigi==2.8.4
lxml==4.3.3
mccabe==0.6.1
pylint==2.3.1
python-box==3.4.0
python-daemon==2.1.2
python-dateutil==2.8.0
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
tornado==4.5.3
typed-ast==1.3.5
urllib3==1.24.2
wrapt==1.11.1
